+++ react_frontend/src/assets/css/chat.css
@@ -56,13 +56,41 @@
     overflow-y: scroll;
 }
 
+.chat_block{
+    display: flex;
+    flex-direction: row;
+    width: 100%;
+    height: 80px;
+    min-height: 80px;
+    align-items: center;
+    justify-content: space-between;
+    padding: 0.5rem 1rem;
+}
+
+.chat_block:hover {
+    background-color: rgb(248,248,248);
+}
+
+
+.chat_badge{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+    width: 1rem;
+    height: 1rem;
+    color: white;
+    background-color: red;
+    border-radius: 50%;
+}
+
 .user_profile {
     display: flex;
     flex-direction: row;
     width: 100%;
     height: 15%;
     justify-content: space-between;
-    padding: 0 1rem;
+    padding: 1rem 1rem;
     align-items: center;
     gap: 1rem;
 }
@@ -72,8 +100,12 @@
     flex-direction: row;
     width: 100%;
     height: 15%;
+    min-height: 15%;
     justify-content: space-between;
     padding: 0.5rem 1rem;
     align-items: center;
-    gap: 1rem;
+}
+
+.user_lists:hover {
+    background-color: rgb(248,248,248);
 }
