+++ react_frontend/src/chat/Chatting.js
@@ -1,3 +1,11 @@
+import React from "react";
+import * as chatStyles from '../assets/css/chat.css'
+
 export default function Chatting(params) {
     
+    return (
+        <div className={chatStyles.chatlist}>
+            hello~
+        </div>
+    );
 };
